+++ src/main/kotlin/racingcar/RacingCarController.kt
@@ -1,7 +1,7 @@
 package racingcar
 
-import racingcar.domain.Car
-import racingcar.domain.Cars
+import racingcar.domain.car.Car
+import racingcar.domain.car.Cars
 import racingcar.domain.numberpicker.RandomNumberPicker
 import racingcar.view.InputView
 import racingcar.view.OutputView
