+++ src/test/kotlin/racingcar/domain/CarsTest.kt
@@ -3,6 +3,8 @@ package racingcar.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
+import racingcar.domain.car.Car
+import racingcar.domain.car.Cars
 import racingcar.domain.numberpicker.NumberPicker
 
 class CarsTest {
